+++ public/classes/models/abstract/Item.js
@@ -1,6 +1,7 @@
 import ItemDom from './ItemDom.js'
 import ItemRepo from './ItemRepo.js'
 import ItemListener from './ItemListener.js'
+import ListerRepo from './ListerRepo.js'
 import LOG from '../../../system/LOG.js'
 
 export default class Item {
@@ -24,6 +25,30 @@ export default class Item {
   get Dom()      { return this._dom   || (this._dom   = new ItemDom(this)) }
   get Repo()     { return this._repo  || (this._repo  = new ItemRepo(this)) }
   get Listener() { return this._listen|| (this._listen= new ItemListener(this)) }
+
+  async enterChildren() {
+    if (!this.parentLister) return
+    const ChildClass = this.parentLister.constructor.CHILD_CLASS
+    if (!ChildClass) return
+    const project_id = this.project_id ?? this.id
+    const child = new ChildClass({
+      id:           this.lister_id ?? this.id,
+      project_id,
+      parentLister: this.parentLister,
+    })
+    await child.Repo.load()
+    if (child._missing) {
+      const result   = await ListerRepo.createLister({ type: `${child.minClass}s`, parent_item_id: this.id, project_id })
+      child.id       = result.id
+      child._missing = false
+      await this.onChildListerCreated?.(child)
+      await child.Repo.load()
+    }
+    this.parentLister.Listener.detach()
+    child.Dom.render()
+    child.Listener.attach(child.Dom.container)
+    if (child.items.length === 0) await child.createNew()
+  }
 
   onEnter() {
     LOG.m(1, 'Item.onEnter', { id: this.id, editing: this.editing })