+++ public/classes/models/abstract/ItemListener.js
@@ -7,9 +7,10 @@ export default class ItemListener extends BaseListener {
   get target() { return this.item }
 
   static LISTENERS = {
-    Enter:  { nokey: 'onEnter' },
-    Escape: { nokey: 'onEscape' },
-    Tab:    { nokey: 'onTab' },
+    Enter:      { nokey: 'onEnter'       },
+    Escape:     { nokey: 'onEscape'      },
+    Tab:        { nokey: 'onTab'         },
+    ArrowRight: { nokey: 'enterChildren' },
   }
 
 }