+++ public/classes/models/core/Event.js
@@ -1,0 +1,7 @@
+import Item from '../abstract/Item.js'
+
+export default class Event extends Item {
+  get PROPS() {
+    return [{ name: 'title', type: 'text' }]
+  }
+}