+++ public/classes/models/core/Project.js
@@ -1,5 +1,5 @@
 import Item from '../abstract/Item.js'
-
+import ListerRepo from '../abstract/ListerRepo.js'
 import {PROJECT_STATES, PROJECT_TYPES} from '../constants/Project.js'
 
 
@@ -10,5 +10,10 @@ export default class Project extends Item {
       , {name: 'state', type: 'select', values: PROJECT_STATES }
       , {name: 'type', type: 'select', values: PROJECT_TYPES}
     ])}
-  
+
+  static async onCreated(id) {
+    const evtLister = await ListerRepo.createLister({ type: 'events', parent_item_id: id, project_id: id })
+    if (evtLister?.id) await ListerRepo.createItem(evtLister.id, { title: 'Acte I' }, { project_id: id })
+  }
+
 }